refactor: [WIP] update data file names
